[Refact] 함수 모듈화, 헤더 생성
