+++ srcs/main.hpp
@@ -1,0 +1,9 @@
+#ifndef _MAIN_HPP_
+#define _MAIN_HPP_
+
+#include "server_socket.hpp"
+#include "client_socket.hpp"
+
+void err_exit(std::string error_msg);
+
+#endif
